FNA Version update, not including new FNA3D or FAUDIO files

+++ FNA/src/Graphics/DxtUtil.cs
@@ -1,6 +1,6 @@
 #region License
 /* FNA - XNA4 Reimplementation for Desktop Platforms
- * Copyright 2009-2022 Ethan Lee and the MonoGame Team
+ * Copyright 2009-2023 Ethan Lee and the MonoGame Team
  *
  * Released under the Microsoft Public License.
  * See LICENSE for details.
@@ -321,7 +321,7 @@ namespace Microsoft.Xna.Framework.Graphics
 			byte alpha0 = imageReader.ReadByte();
 			byte alpha1 = imageReader.ReadByte();
 
-			ulong alphaMask = imageReader.ReadByte();
+			ulong alphaMask = (ulong) imageReader.ReadByte();
 			alphaMask += (ulong) imageReader.ReadByte() << 8;
 			alphaMask += (ulong) imageReader.ReadByte() << 16;
 			alphaMask += (ulong) imageReader.ReadByte() << 24;
